1.divide relative test functions into individual files; 2.incorporate behave into project
