+++ tests/end_to_end/features/chain.feature
@@ -1,0 +1,6 @@
+Feature: enumerate all possible cases
+
+  Scenario: we want to enumerate all possible cases
+    Given: we have known all possible cases related to known signals
+    When: we are provided some signals
+    Then: chain will give us the cases correctly
